fix Edit and Delete functionality

+++ ImageHoster/src/main/java/ImageHoster/controller/ImageController.java
@@ -94,13 +94,21 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, HttpSession session,Model model) {
+        String error = "Only the owner of the image can edit the image";
         Image image = imageService.getImage(imageId);
-
         String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
         model.addAttribute("tags", tags);
-        return "images/edit";
+        model.addAttribute("editError", error);
+        User user = (User) session.getAttribute("loggeduser");
+        if(user.getId()==image.getUser().getId()){
+            return "images/edit";
+        }
+        else{
+            return "images/image";
+        }
+
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -134,7 +142,8 @@ public class ImageController {
         updatedImage.setDate(new Date());
 
         imageService.updateImage(updatedImage);
-        return "redirect:/images/" + updatedImage.getTitle();
+       // return "redirect:/images/" + updatedImage.getTitle();
+        return "images/image" ;
     }
 
 
@@ -142,9 +151,25 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId,HttpSession session,Model model) {
+        String error = "Only the owner of the image can edit the image";
+        Image image = imageService.getImage(imageId);
+        String tags = convertTagsToString(image.getTags());
+        model.addAttribute("image", image);
+        model.addAttribute("tags", tags);
+        model.addAttribute("deleteError", error);
+        User user = (User) session.getAttribute("loggeduser");
+        if(user.getId()==image.getUser().getId()){
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
+        else{
+            return "images/image";
+        }
+
+
+
+
     }
